Moved separate form validation functions into one file
